JSONUsernamePasswordAuthenticationFilter moved to new package - ~~.web
